Fixed Issues with Blobs
There was an issue with the matrix solver that was fixed since we had the wrong system of equations. If we tile a 2 by 2, the horizontal/vertical blobs will get counted 6 times and not 3. 

Also, we condensed the code and fixed an issue that was causing the blob detector not to catch some aggregations. Thus, we added a new function in order to do this.