1D Code has been added
This is the initial start for the 1D code. We have included a worksheet that runs all of the relevant code with all of the modules that allow for the different deposition functions, the plotting mechanisms, and the module that solves and updates the evolution of the PDE we are solving.