Updated PDE Solving Mechanisms
We have fixed a bug that did not allow for the correct solving of these equations. The two methods coincide with each other and we have added some stability analysis as well for the worksheet, so that the user can see if the parameters given induce a stable or unstable constant steady state. All of the timing should match up, but now it is necessary to check convergence temporally and spatially.